% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\docType{methods}
\name{addTrees,honestRF-method}
\alias{addTrees,honestRF-method}
\alias{addTrees,}
\alias{honestRF-method}
\title{addTrees-honestRF}
\usage{
\S4method{addTrees}{honestRF}(object, ntree)
}
\value{
A `honestRF` object
}
