% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\name{addTrees-honestRF}
\alias{addTrees-honestRF}
\title{addTrees-honestRF}
\usage{
addTrees(object, ntree)
}
\arguments{
\item{object}{A `honestRF` object.}

\item{ntree}{Number of new trees to add}
}
\description{
Add more trees to the existing forest.
}
