% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\docType{data}
\name{autohonestRF-honestRF}
\alias{autohonestRF-honestRF}
\title{autohonestRF-honestRF}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{sampsize}{The size of total samples to draw for the training data.}

\item{num_iter}{Maximum iterations/epochs per configuration. Default is 1024.}

\item{eta}{Downsampling rate. Default value is 2.}

\item{verbose}{if tuning process in verbose mode}

\item{seed}{random seed}

\item{nthread}{Number of threads to train and predict thre forest. The
default number is 0 which represents using all cores.}
}
\description{
Autotune a honestRF based on the input dataset. The methodology
is based on paper `Hyperband: A Novel Bandit-Based Approach to
Hyperparameter Optimization` by Lisha Li, et al.
}
\keyword{datasets}
