% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\name{autohonestRF}
\alias{autohonestRF}
\alias{autohonestRF,}
\alias{honestRF-method}
\title{autohonestRF-honestRF}
\usage{
autohonestRF(x, y, sampsize = as.integer(nrow(x) * 0.75), num_iter = 1024,
  eta = 2, verbose = FALSE, seed = 24750371, nthread = 0)
}
\value{
A `honestRF` object
}
