% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune_hyperband2.R
\name{autoJointHonestRF}
\alias{autoJointHonestRF}
\title{Autotuning for X-Learner with honest RF for both stages}
\usage{
autoJointHonestRF(feat, tr, yobs, sample.fraction = 0.75, num_iter = 3^8,
  eta = 3, verbose = TRUE, seed = 24750371, nthread = 0,
  tune_iter = 10)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{num_iter}{number of iterations.}

\item{eta}{..}

\item{verbose}{..}

\item{seed}{..}

\item{nthread}{..}

\item{sampsize}{..}

\item{firststageVar}{..}

\item{secondstageVar}{..}
}
\description{
[TBD]
}
\seealso{
\code{\link{X_RF_autotune_simple}}, \code{\link{X_RF_autotune_gpp}},
}
