% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\docType{methods}
\name{CateBIAS-Meta-learner}
\alias{CateBIAS-Meta-learner}
\alias{CateBIAS,Meta-learner-method}
\title{CateBIAS-Meta-learner}
\usage{
\S4method{CateBIAS}{`CATE-estimator`}(theObject, feature_new,
  method = "maintain_group_ratios", B = 2000, nthread = 0,
  verbose = TRUE)
}
\value{
A data frame of estimated CATE bias
}
\description{
Return the estimated bias for the CATE
}
