% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{CateBIAS}
\alias{CateBIAS}
\alias{CateBIAS,}
\alias{Meta-learner-method}
\title{CateBIAS-Meta-learner}
\usage{
CateBIAS(theObject, feature_new, method = "maintain_group_ratios",
  B = 2000, nthread = 0, verbose = TRUE)
}
\value{
A data frame of estimated bias
}
\description{
Return the estimated bias for the CATE
}
