% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\docType{methods}
\name{CateCI-Meta-learner}
\alias{CateCI-Meta-learner}
\alias{CateCI,Meta-learner-method}
\title{CateCI-Meta-learner}
\usage{
\S4method{CateCI}{`Meta-learner`}(theObject, feature_new,
  method = "maintain_group_ratios", bootstrapVersion = "normalApprox",
  B = 2000, nthread = 0, verbose = TRUE)
}
\arguments{
\item{theObject}{A `Meta-learner` object}

\item{feature_new}{A feature data frame}

\item{method}{Different versions of the bootstrap.}

\item{bootstrapVersion}{default is normalApprox which will just use the 
bootstrap normal approximation to get CI. smoothed will use use CI around the
smoothed bootstrap as introduced by Efron 2014.}

\item{B}{Number of bootstrap samples}

\item{nthread}{number of threads to be used in parallel}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A data frame of estimated CATE Confidence Intervals
}
\description{
Return the estimated confidence intervals for the CATE
}
