% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{CateCI,CATEestimator-method}
\alias{CateCI,CATEestimator-method}
\title{CateCI-MetaLearner}
\usage{
\S4method{CateCI}{CATEestimator}(theObject, feature_new, method, bootstrapVersion, B, nthread, verbose)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}

\item{method}{Different versions of the bootstrap.}

\item{bootstrapVersion}{Default is normalApprox, which will use the 
bootstrap normal approximation to get CI. Smoothed will use CI around the
smoothed bootstrap as introduced by Efron 2014. The third option is to use 
the doubleBootstrap option, which uses a double level bootstrap to calibrate 
the quantiles used in the bootstrap estimation of the intervals. For reference
see https://arxiv.org/pdf/1511.00273.pdf, although this is an older algorithm
which was introduced much earlier.}

\item{B}{Number of bootstrap samples.}

\item{nthread}{Number of threads to be used in parallel.}

\item{verbose}{TRUE for detailed output, FALSE for no output.}
}
\value{
A data frame of estimated CATE confidence intervals.
}
\description{
Returns the estimated confidence intervals for the CATE.
}
\examples{
\dontrun{
require(causalToolbox)

# create example data set
simulated_experiment <- simulate_causal_experiment(
  ntrain = 1000,
  ntest = 1000,
  dim = 10
)
feat <- simulated_experiment$feat_tr
tr <- simulated_experiment$W_tr
yobs <- simulated_experiment$Yobs_tr
feature_test <- simulated_experiment$feat_te

# create the CATE estimator using Random Forests (RF)
xl_rf <- X_RF(feat = feat, tr = tr, yobs = yobs)
CateCI(xl_rf, feature_test, B = 500)
}
}
