% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBART.R
\docType{methods}
\name{CateCI,S_BART-method}
\alias{CateCI,S_BART-method}
\title{CateCI-S_BART}
\usage{
\S4method{CateCI}{S_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{theObject}{A `S_BART` object.}

\item{feature_new}{A feature data frame.}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
