% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBART.R
\docType{methods}
\name{CateCI-T_BART}
\alias{CateCI-T_BART}
\alias{CateCI,T_BART-method}
\title{CateCI-T_BART}
\usage{
\S4method{CateCI}{T_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{theObject}{A `T_BART` object.}

\item{feature_new}{A feature data frame}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
