% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\docType{methods}
\name{CateCI-X_BART}
\alias{CateCI-X_BART}
\alias{CateCI,X_BART-method}
\title{CateCI-X_BART}
\usage{
\S4method{CateCI}{X_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{theObject}{A `X_BART` object.}

\item{feature_new}{A data frame.}

\item{verbose}{TRUE for detailed output FALSE for no output. 
(Should the training output be posted?)}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
