% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\docType{methods}
\name{CateCI-X_hRF}
\alias{CateCI-X_hRF}
\alias{CateCI,}
\alias{X_hRF-method}
\title{CateCI-X_hRF}
\usage{
\S4method{CateCI}{X_RF}(theObject, feature_new, method = "n2TBS", B,
  nthread = 8)
}
\arguments{
\item{feature_new}{A data frame.}

\item{method}{different versions of the bootstrap. Only n2TBS implemented}

\item{B}{number of bootstrap samples.}

\item{nthread}{number of threats used.}

\item{object}{A `X_hRF` object.}
}
\value{
A data frame of estimated CATE Confidence Intervals
}
\description{
Return the estimated confidence intervals for the CATE
}
