% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\docType{methods}
\name{EstimateAllSampleStatistics-Meta-learner}
\alias{EstimateAllSampleStatistics-Meta-learner}
\alias{EstimateAllSampleStatistics,Meta-learner-method}
\title{EstimateAllSampleStatistics-Meta-learner}
\usage{
\S4method{EstimateAllSampleStatistics}{`Meta-learner`}(theObject, method, B,
  nthread, verbose)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{method}{...}

\item{B}{...}

\item{nthread}{number of threads to be used}

\item{verbose}{...}
}
\description{
...
}
