% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBART.R
\docType{methods}
\name{EstimateAllSampleStatistics,S_BART-method}
\alias{EstimateAllSampleStatistics,S_BART-method}
\title{EstimateAllSampleStatistics-S_BART}
\usage{
\S4method{EstimateAllSampleStatistics}{S_BART}(theObject,
  verbose = FALSE)
}
\arguments{
\item{theObject}{A `S_BART` object}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\description{
Return the estimated CATE
}
