% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBART.R
\docType{methods}
\name{EstimateAllSampleStatistics-T_BART}
\alias{EstimateAllSampleStatistics-T_BART}
\alias{EstimateAllSampleStatistics,T_BART-method}
\title{EstimateAllSampleStatistics-T_BART}
\usage{
\S4method{EstimateAllSampleStatistics}{T_BART}(theObject,
  verbose = FALSE)
}
\arguments{
\item{theObject}{A "T_BART" object}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\description{
Return the estimated CATE
}
