% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{EstimateAllSampleStatistics}
\alias{EstimateAllSampleStatistics}
\title{Method EstimateAllSampleStatistics}
\usage{
EstimateAllSampleStatistics(theObject, feature_new,
  method = "maintain_group_ratios", B = 200, nthread = 0,
  verbose = TRUE, ...)
}
\description{
Method EstimateAllSampleStatistics
}
