% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRRF.R
\name{EstimateCate,DR_RF-method}
\alias{EstimateCate,DR_RF-method}
\title{EstimateCate-DR_hRF}
\usage{
\S4method{EstimateCate}{DR_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}
}
\description{
Returns the estimated CATE.
}
