% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOBART.R
\docType{methods}
\name{EstimateCate-MO_BART}
\alias{EstimateCate-MO_BART}
\alias{EstimateCate,MO_BART-method}
\title{EstimateCate-MO_BART}
\usage{
\S4method{EstimateCate}{MO_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)
}
\arguments{
\item{theObject}{A `MO_BART` object.}

\item{feature_new}{A data frame.}

\item{verbose}{TRUE for detailed output FALSE for no output}

\item{return_CI}{TRUE for predictions and confidence intervals FALSE for only
predictions.}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
