% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrf.R
\docType{methods}
\name{EstimateCate-S_hRF}
\alias{EstimateCate,}
\alias{EstimateCate-S_hRF}
\alias{S_hRF-method}
\title{EstimateCate-S_hRF}
\usage{
\S4method{EstimateCate}{S_RF}(theObject, feature_new)
}
\arguments{
\item{feature_new}{A data frame.}

\item{object}{A `S_hRF` object.}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}

