% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrf.R
\docType{methods}
\name{EstimateCate-S_RF}
\alias{EstimateCate-S_RF}
\alias{EstimateCate,S_RF-method}
\title{EstimateCate-S_hRF}
\usage{
\S4method{EstimateCate}{S_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `S_hRF` object.}

\item{feature_new}{A data frame.}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
