% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thrf.R
\docType{methods}
\name{EstimateCate-T_hRF}
\alias{EstimateCate-T_hRF}
\alias{EstimateCate,}
\alias{T_hRF-method}
\title{EstimateCate-T_hRF}
\usage{
\S4method{EstimateCate}{T_RF}(theObject, feature_new)
}
\arguments{
\item{feature_new}{A data frame.}

\item{object}{A `T_hRF` object.}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
