% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRF.R
\docType{methods}
\name{EstimateCate-T_RF}
\alias{EstimateCate-T_RF}
\alias{EstimateCate,T_RF-method}
\title{EstimateCate-T_hRF}
\usage{
\S4method{EstimateCate}{T_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `T_hRF` object.}

\item{feature_new}{A feature data frame.}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
