% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\docType{methods}
\name{EstimateCate-X_BART}
\alias{EstimateCate-X_BART}
\alias{EstimateCate,X_BART-method}
\title{EstimateCate-X_BART}
\usage{
\S4method{EstimateCate}{X_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)
}
\arguments{
\item{theObject}{A `X_BART` object.}

\item{feature_new}{A data frame.}

\item{verbose}{TRUE for detailed output FALSE for no output}

\item{return_CI}{...}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
