% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thrf.R, R/Xhrf.R
\docType{methods}
\name{EstimateCate-X_hRF}
\alias{EstimateCate-X_hRF}
\alias{EstimateCate,}
\alias{X_hRF-method}
\alias{EstimateCate-X_hRF}
\alias{EstimateCate,}
\alias{X_hRF-method}
\title{EstimateCate-X_hRF}
\usage{
\S4method{EstimateCate}{T_RF}(theObject, feature_new)

\S4method{EstimateCate}{X_RF}(theObject, feature_new)
}
\arguments{
\item{feature_new}{A data frame.}

\item{object}{A `X_hRF` object.}

\item{object}{A `X_hRF` object.}

\item{feature_new}{A data frame.}
}
\value{
A vector of predicted CATE

A vector of predicted CATE
}
\description{
Return the estimated CATE

Return the estimated CATE
}
