% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\docType{methods}
\name{EstimateCate-X_RF}
\alias{EstimateCate-X_RF}
\alias{EstimateCate,X_RF-method}
\title{EstimateCate-X_hRF}
\usage{
\S4method{EstimateCate}{X_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `X_hRF` object.}

\item{feature_new}{A data frame.}
}
\value{
A vector of predicted CATE
}
\description{
Return the estimated CATE
}
