% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R, R/SBART.R, R/XRF.R, R/TRF.R
\docType{methods}
\name{EstimateCate}
\alias{EstimateCate}
\alias{EstimateCate-S_BART}
\alias{EstimateCate-X_RF}
\alias{EstimateCate-T_RF}
\title{Method EstimateCate}
\usage{
EstimateCate(theObject, feature_new, ...)

\S4method{EstimateCate}{S_BART}(theObject, feature_new, verbose = FALSE)

\S4method{EstimateCate}{X_RF}(theObject, feature_new)

\S4method{EstimateCate}{T_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{feature_new}{A feature data frame.}

\item{...}{Additional parameters that are specific for some meta-learners}
}
\description{
Returns the estimated CATE.
}
