% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\docType{methods}
\name{getOOB,honestRF-method}
\alias{getOOB,}
\alias{getOOB,honestRF-method}
\alias{honestRF-method}
\title{getOOB-honestRF}
\usage{
\S4method{getOOB}{honestRF}(object, noWarning = FALSE)
}
\value{
The OOB error of the forest.
}

