% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_double_robust.R
\name{gof_double_robust}
\alias{gof_double_robust}
\title{gof_double_robust}
\usage{
gof_double_robust(feat, yobs, tr, estimator, emin, k = 5, verbose = FALSE)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observations}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{emin}{Minimum value of the propensity score}

\item{k}{k fold cross validation}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
error
}
