% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF.R
\name{GoF_estimator}
\alias{GoF_estimator}
\title{GoF_estimator}
\usage{
GoF_estimator(feat, tr, yobs)
}
\arguments{
\item{feat}{features in data.frame.}

\item{tr}{treatment assignment, 0 for control and 1 for treated.}

\item{yobs}{the observed outcome.}
}
\value{
A `GoF_estimator` object.
}
\description{
Out-of-sample Goodness-of-fit measure
}
