% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_fishers_exact.R
\name{gof_fisher_exact}
\alias{gof_fisher_exact}
\title{gof_fisher_exact}
\usage{
gof_fisher_exact(feat, yobs, tr, estimator, k = 2, verbose = FALSE,
  loss.type = c("min", "mean", "0", "1", "wilcox", "ks"))
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observed values}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{learner constructor}

\item{k}{number of folds used for cross-validation}

\item{verbose}{logical. If TRUE then detailed updates will be printed}

\item{loss.type}{One of 'min', 'mean', '0', '1', 'wilcox', and 'ks'. 
Determines how the loss should be computed from the pvalues of the ks and 
wilcoxon rank tests.}
}
\value{
mean(error) and sd(error)
}
