% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_matched.R
\name{gof_matched}
\alias{gof_matched}
\title{gof_matched}
\usage{
gof_matched(feat, yobs, tr, estimator, estimand, k = 2, emin = 1e-05)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observations}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{estimand}{ATE, ATT or ATC}

\item{k}{we are doing a k fold cross validation}

\item{emin}{the pscore prediciton will be bounded between emin and 1 - emin
to avoid decide by 0 error}
}
\value{
mean(error) and sd(error)
}
\description{
gof_matched
}
