% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_matching.R
\name{gof_matching}
\alias{gof_matching}
\title{gof_matching}
\usage{
gof_matching(feat, yobs, tr, estimator, estimand = "ATT", k = 2,
  replace = TRUE, emin = 1e-05, verbose = FALSE)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observations}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{estimand}{ATE, ATT or ATC}

\item{k}{Number of folds used for cross validation}

\item{replace}{Whether or not to replace samples while matching}

\item{emin}{Minimum value of the propensity score}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
mean(error) and sd(error)
}
\description{
Match each unit in one group(treatment/control) to units in
another group(treatment/control) based on the chosen estimand. For each iteration
of cross validation, assign treatment lablels (0 or 1) with probability given 
by averaged propensity score for each pair of match and estimate CATE. Return
the mean squared error and its standard deviation of CATE from the true difference.
}
