% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_subset.R
\name{gof_subset}
\alias{gof_subset}
\title{gof_subset}
\usage{
gof_subset(feat, yobs, tr, estimator, important.features = colnames(feat),
  min.treat.size.per.group = 25, normalize = "studentize", k = 5)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observations}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{important.features}{names of features which should be used in 
quickmatch to find the relevant subgroups. Only features specified here will 
be used to create the subgroups}

\item{min.treat.size.per.group}{the minimum size of each treatment group 
per matched subset. For example, if it is chosen to be 25 (the default), then 
at least 25 units in each subgroup are treated and 25 are in the control
goup}

\item{normalize}{Specifies how the distance should be normalized. 
"none": no normalization. 
"mahalanobize": normalization by var(data)
"studentize" (default): normalization is done with the diagonal of var(data)}

\item{k}{k fold cross validation}

\item{min.n.group}{Minimum number of groups}
}
\value{
mean(error) and sd(error)
}
