% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF.R
\docType{methods}
\name{GoF_TOT-GoF_estimator}
\alias{GoF_TOT-GoF_estimator}
\title{GoF_TOT-GoF_estimator}
\usage{
\S4method{GoF_TOT}{GoF_estimator}(theObject, learner_func)
}
\arguments{
\item{learner_func}{The CATE estimator to be evaluated.}

\item{object}{A `GoF_estimator` object.}
}
\value{
The goodness-of-fit measure for the estimator.
}
\description{
TOT method for measuring goodnes-of-fit
}
