% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_transformed.R
\name{gof_transformed}
\alias{gof_transformed}
\title{gof_transformed}
\usage{
gof_transformed(feat, yobs, tr, estimator, k = 2, emin = 1e-05,
  verbose = FALSE)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observed values}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{learner constructor}

\item{k}{number of folds used for cross-validation}

\item{emin}{the pscore estimate will be bounded between emin and 1 - emin
in order to avoid any divide-by-zero errors}

\item{verbose}{logical. If TRUE then detailed updates will be printed}
}
\value{
mean(error) and sd(error)
}
\description{
gof_transformed
}
