% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_transformed.R
\docType{methods}
\name{GoF_yStar}
\alias{GoF_yStar}
\title{GoF_yStar}
\usage{
\S4method{GoF_yStar}{ANY}(feat, yobs, tr, estimator, seed)
}
\arguments{
\item{feat}{a data frame of features}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{obs}{a vector of observations}
}
\description{
GoF_yStar
}
