% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF_R.R
\docType{class}
\name{honestRF_R-class}
\alias{honestRF_R-class}
\title{honestRF_R Constructor}
\description{
`honestRF_R` inherits `RF`, which serves as a modified version of
`RF`. The major change is that it trains honest trees instead of adaptive
trees. Adaptive trees are the original trees as they were used in the
original implementation and honest trees differ in that they require two
data sets to do tree estimation. One data set is used to create the trees
and the other one is used to receive the leaf estimates.
}
\section{Slots}{

\describe{
\item{\code{splitratio}}{Proportion of the training data used as the splitting
dataset. It is a ratio between 0 and 1. If the ratio is 1, then essentially
splitting dataset becomes the total entire sampled set and the averaging
dataset is empty. If the ratio is 0, then the splitting data set is empty
and all the data is used for the averaging data set (This is not a good
usage however since there will be no data available for splitting).}
}}

