% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRFRcpp.R
\docType{class}
\name{honestRFRcpp-class}
\alias{honestRFRcpp-class}
\title{honestRFRcpp-Constructor}
\description{
`honestRFRcpp` object implementing the most basic version of
a random forest.
}
\section{Slots}{

\describe{
\item{\code{forest}}{A list of `RFTree` objects in the forest. If the class is
extended, the list may contain the corresponding extended `RFTree` object.}

\item{\code{categoricalFeatureCols}}{A list of index for all categorical data. Used
for trees to detect categorical columns.}

\item{\code{categoricalFeatureMapping}}{A list of encoding details for each
categorical column, including all unique factor values and their
corresponding numeric representation.}
}}

