% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRFTree.R
\docType{class}
\name{honestRFTree-class}
\alias{honestRFTree}
\alias{honestRFTree-class}
\title{honstRFTree Constructor}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{mtry}{The number of variables randomly selected at each split point.
The default value is set to be one third of total number of features of the
training data.}

\item{nodesize}{The minimum observations contained in terminal nodes. The
default value is 5.}

\item{sampleIndex}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}

\item{splitrule}{A string to specify how to find the best split among all
candidate feature values. The current version only supports `variance` which
minimizes the overall MSE after splitting. The default value is `variance`.}

\item{categoricalFeatureCols}{A list of index for all categorical data. Used
for trees to detect categorical columns.}
}
\description{
`honestRFTree` inherits `RFTree`, which serves as a modified
version of `RFTree`. The major change is that when `honestRFTree` determines
a valid split point, it looks at whether or not it is valid for both
averaging and splitting dataset.

Create a honestRFTree by making specifc observatios as splitting
and averaging dataset.
}
\keyword{datasets}

