% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRFTree.R
\name{honestRFTree}
\alias{honestRFTree}
\alias{honestRFTree,}
\alias{honestRFTree-method}
\title{honstRFTree Constructor}
\usage{
honestRFTree(x, y, mtry = max(floor(ncol(x)/3), 1),
  nodesize = list(averagingNodeSize = 5, splittingNodeSize = 5),
  sampleIndex = list(averagingSampleIndex = 1:length(y), splittingSampleIndex
  = 1:length(y)), splitrule = "variance")
}
\value{
a `honestRFTree` object
}
