% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRFRcpp.R
\docType{methods}
\name{predict-honestRFRcpp}
\alias{predict-honestRFRcpp}
\alias{predict,}
\alias{honestRFRcpp-method}
\title{predict-honestRFRcpp}
\usage{
\S4method{predict}{honestRFRcpp}(object, feature.new)
}
\arguments{
\item{object}{A `honestRFRcpp` object.}

\item{feature.new}{A data frame of testing predictors.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
