% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\docType{methods}
\name{predict-RFTree}
\alias{predict-RFTree}
\alias{predict,}
\alias{RFTree-method}
\title{predict-RFTree}
\usage{
\S4method{predict}{RFTree}(object, feature.new, x, y, avgfunc)
}
\arguments{
\item{object}{RFTree object.}

\item{feature.new}{A data frame or a matrix of all testing predictors.}

\item{x}{A data frame or a matrix of all training predictors.}

\item{y}{A vector of all training responses.}

\item{avgfunc}{An averaging function to average all the input data. The input
of this function should be a dataframe of predictors `x` and a vector of
outcome `y`. The output is a scalar. The default is to take the mean of all
the `y`s.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the current tree.
}
