% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{preprocess_testing}
\alias{preprocess_testing}
\title{preprocess_testing}
\usage{
preprocess_testing(x, categoricalFeatureCols, categoricalFeatureMapping)
}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{categoricalFeatureCols}{A list of index for all categorical data. Used
for trees to detect categorical columns.}

\item{categoricalFeatureMapping}{A list of encoding details for each
categorical column, including all unique factor values and their
corresponding numeric representation.}
}
\value{
A preprocessed training dataaset x
}
\description{
Perform preprocessing for the testing data, including
converting data to dataframe, and testing if the columns are consistent
with the training data and encoding categorical data into numerical
representation in the same way as training data.
}

