% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{preprocess_training}
\alias{preprocess_training}
\title{preprocess_training}
\usage{
preprocess_training(x, y)
}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}
}
\value{
A list of two datasets along with necessary information that
encoding the preprocessing.
}
\description{
Perform preprocessing for the training data, including
converting data to dataframe, and encoding categorical data into numerical
representation.
}

