% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{preprocessing_data}
\alias{preprocessing_data}
\title{Preprocess Data}
\usage{
preprocessing_data(x, labels = NULL)
}
\arguments{
\item{x}{A data frame of all predictors.}

\item{labels}{A list contains all encodings that are used to encode
categorical columns in the training data. When the function is called for
training dataset, the labels should be set to NULL. When the function is
called for testing dataset, the labels must be provided in order to
successfully convert the data into the format that is consistent to training
data.}
}
\value{
A list of the transformed object `x`, and encoding information
`labels`.
}
\description{
A simple preprocessing function to convert `x` to dataframe,
and also convert categorical data into one-hot encoding.
}
