% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFNode.R
\docType{methods}
\name{printnode-RFNode}
\alias{RFNode-method}
\alias{printnode,}
\alias{printnode,RFNode-method}
\alias{printnode-RFNode}
\title{printnode-RFNode}
\usage{
printnode(object, indentSpace = 0)

\S4method{printnode}{RFNode}(object, indentSpace = 0)
}
\arguments{
\item{object}{RFNode object}

\item{indentSpace}{The indentation space of the current node in the standard
output. The initial value is 0. Once the tree grows, it increments by 2.}
}
\description{
Print the type of the node and its prediction. If it is a tree
node, it will recursively print all its children.
}

