% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\name{recursivePartition-RFTree}
\alias{recursivePartition-RFTree}
\alias{recursivePartition}
\title{recursivePartition-RFTree}
\usage{
recursivePartition(x, y, mtry = max(floor(ncol(x)/3), 1),
  sampleIndex = list(averagingSampleIndex = 1:length(y), splittingSampleIndex
  = 1:length(y)), nodesize = list(splittingNodeSize = 5, averagingNodeSize =
  5), splitrule = "variance")
}
\arguments{
\item{x}{A data frame or a matrix of all training predictors.}

\item{y}{A vector of all training responses.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
The default value is set to be one third of total feature amount.}

\item{sampleIndex}{A list of index of dataset used in this node and its
children. `sampleIndex` contains two keys `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `splittingSampleIndex` is the same as `averagingSampleIndex`.}

\item{nodesize}{Minimum size of terminal nodes. Setting this number larger
causes smaller trees to be grown (and thus take less time). This parameter is
actually a list containing the values for both `splittingNodeSize` and
`averagingNodeSize`.}

\item{splitrule}{A splitting rule to determine the best split point among the
features. There are two possible split rules in the package, `variance` and
`maxstats`. The default is `variance` to minimize the overall MSE.}
}
\value{
root node.
}
\description{
Grow the decision tree by recursively finding the best split
feature and value.
}
