% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\docType{class}
\name{RF-class}
\alias{RF-class}
\alias{RF}
\title{`RF` Constructor}
\usage{
RF(x, y, ntree = 500, replace = TRUE, sampsize = if (replace) nrow(x) else
  ceiling(0.632 * nrow(x)), mtry = max(floor(ncol(x)/3), 1), nodesize = 5,
  nthread = 1, splitrule = "variance", avgfunc = avgMean)
}
\arguments{
\item{x}{A data frame or a matrix of all training predictors.}

\item{y}{A vector of all training responses.}

\item{ntree}{Number of trees to grow. This should not be set to too small to
ensure that every input row gets predicted at least a few times. The default
value is 500.}

\item{replace}{Indicator of whether sampling of cases be done with or without
replacement.}

\item{sampsize}{Size(s) of sample to draw.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
The default value is set to be one third of total feature amount.}

\item{nodesize}{Minimum size of terminal nodes. Setting this number larger
causes smaller trees to be grown (and thus take less time). The default value
is 5.}

\item{nthread}{Number of threads to use in parallel}

\item{splitrule}{A splitting rule to determine the best split point among the
features. There are two possible split rules in the package, `variance` and
`maxstats`. The default is `variance` to minimize the overall MSE.}

\item{avgfunc}{An averaging function to average all the input data. The input
of this function should be a dataframe of predictors `x` and a vector of
outcome `y`. The output is a scalar. The default is to take the mean of all
the `y`s.}
}
\value{
A `RF` object.
}
\description{
`RF` object is an object implementing the most basic version of
a random forest.

`RF` Constructor
}
\section{Slots}{

\describe{
\item{\code{x}}{A data frame or a matrix of all training predictors.}

\item{\code{y}}{A vector of all training responses.}

\item{\code{ntree}}{Number of trees to grow. This should not be set to too small to
ensure that every input row gets predicted at least a few times. The default
value is 500.}

\item{\code{replace}}{Indicator of whether sampling of cases be done with or without
replacement.}

\item{\code{sampsize}}{Size(s) of sample to draw.}

\item{\code{mtry}}{Number of variables randomly sampled as candidates at each split.
The default value is set to be one third of total feature amount.}

\item{\code{nodesize}}{Minimum size of terminal nodes. Setting this number larger
causes smaller trees to be grown (and thus take less time). The default value
is 5.}

\item{\code{splitrule}}{A splitting rule to determine the best split point among the
features. There are two possible split rules in the package, `variance` and
`maxstats`. The default is `variance` to minimize the overall MSE.}

\item{\code{avgfunc}}{An averaging function to average all the input data. The input
of this function should be a dataframe of predictors `x` and a vector of
outcome `y`. The output is a scalar. The default is to take the mean of all
the `y`s.}

\item{\code{forest}}{A list of `RFTree` in the forest. If the class is extended, the
list may contain the corresponding extended `RFTree` object.}
}}

