% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFNode.R
\docType{class}
\name{RFNode-class}
\alias{RFNode-class}
\alias{RFNode}
\title{RFNode Constructor}
\usage{
RFNode(sampleIndex = list(averagingSampleIndex = vector(),
  splittingSampleIndex = vector()), splitFeature = numeric(),
  splitValue = numeric(), child = list())
}
\arguments{
\item{sampleIndex}{A list of index of dataset used in this node and its
children. `sampleIndex` contains two keys `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `splittingSampleIndex` is the same as `averagingSampleIndex`.}

\item{splitFeature}{Name of the feature that is used for splitting in this
node.}

\item{splitValue}{The value that is used for splitting in this node.}

\item{child}{If the node is not a leaf node, the `child` will be a list of
two `RFNode` objects. If it is a leaf node, the `child` will be `NULL`.}
}
\value{
a `RFNode` object
}
\description{
`RFNode` is the basic element inside a `RFTree`. For each node,
it contains the corresponding data that are assigned to the node. The
`RFNode` can be either a leaf or a tree node (non-leaf). If it is a leaf
node, `avgfunc` can be called to aggregate all the observations in the node
and save a prediction for the node. If it is a tree node, it will contains
`leftChild` and `rightChild` which will be another two `RFNode`s. The parent
node can be used to track the `splitFeature` and `splitValue`. Please note
that in the `RFNode`, `sampleIndex` essentially tells what data reside
in the node. Every node contains two `sampleIndex`: `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `RFNode` has the same `averagingSampleIndex` and
`splittingSampleIndex`.

RFNode Constructor
}
\section{Slots}{

\describe{
\item{\code{sampleIndex}}{A list of index of dataset used in this node and its
children. `sampleIndex` contains two keys `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `splittingSampleIndex` is the same as `averagingSampleIndex`.}

\item{\code{splitFeature}}{Name of the feature that is used for splitting in this
node.}

\item{\code{splitValue}}{The value that is used for splitting in this node.}

\item{\code{child}}{If the node is not a leaf node, the `child` will be a list of two
`RFNode` objects. If it is a leaf node, the `child` will be `NULL`.}

\item{\code{nSplit}}{Number of observations in the splitting dataset in this node.}

\item{\code{nAverage}}{Number of observations in the averaging dataset in this node.}
}}

