% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFNode.R
\name{RFNode}
\alias{RFNode}
\alias{RFNode,}
\alias{RFNode-method}
\title{RFNode Constructor}
\usage{
RFNode(sampleIndex = list(averagingSampleIndex = vector(),
  splittingSampleIndex = vector()), splitFeature = numeric(),
  splitValue = numeric(), child = list())
}
\value{
a `RFNode` object
}

