% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\docType{class}
\name{RFTree-class}
\alias{RFTree-class}
\alias{RFTree}
\title{`RFTree` Constructor}
\usage{
RFTree(x, y, mtry = max(floor(ncol(x)/3), 1), nodesize = 5,
  sampleIndex = 1:length(y), splitrule = "variance")
}
\arguments{
\item{x}{A data frame or a matrix of all training predictors.}

\item{y}{A vector of all training responses.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
The default value is set to be one third of total feature amount.}

\item{nodesize}{Minimum size of terminal nodes. Setting this number larger
causes smaller trees to be grown (and thus take less time). The default value
is 5.}

\item{sampleIndex}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}

\item{splitrule}{A splitting rule to determine the best split point among the
features. There are two possible split rules in the package, `variance` and
`maxstats`. The default is `variance` to minimize the overall MSE.}
}
\value{
a `RFTree` object
}
\description{
`RFTree` is the unit component in the `RF` which composes
`RFNode`. The tree uses recursively partitioning to determine the best
`splitFeature` and `splitValue` for each level, and recursively split the
dataset until it reaches the limitation according to `nodesize`.

The actual storation of `RFTree` is different from its input.
`sampleIndex` and `nodesize` are actually two lists composing of those in
splitting dataset and averaging dataset. In the original random forest,
those are the same value. For other variant of random forest, such as honest
random forest, they can be set differently. The constructor can be
overwritten to reflect the change.
}
\section{Slots}{

\describe{
\item{\code{sampleIndex}}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}

\item{\code{root}}{A `RFNode` object which is the root of the tree. If the class is
extended, the list may contain the corresponding extended `RFNode` object.}
}}

