% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\name{RFTree}
\alias{RFTree}
\alias{RFTree,}
\alias{RFTree-method}
\title{RFTree-Constructor}
\usage{
RFTree(x, y, se = NULL, mtry = max(floor(ncol(x)/3), 1), nodesize = 5,
  sampleIndex = 1:length(y), splitrule = "variance",
  categoricalFeatureCols = list())
}
\value{
a `RFTree` object
}
