% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRF.R
\docType{class}
\name{S_RF-class}
\alias{S_RF-class}
\title{ShRF constructor}
\description{
The `S_RF` object is a S-learner combined with honest random
forest used for both response functions
}
\section{Slots}{

\describe{
\item{\code{feature_train}}{A data frame of all training features.}

\item{\code{tr_train}}{A vector containing 0 for control and 1 for treated variables.}

\item{\code{yobs_train}}{A vector containing the observed outcomes.}

\item{\code{forest}}{A forest object}

\item{\code{creator}}{A function which creates a S_RF}
}}

