% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\name{selectBestFeature-RFTree}
\alias{selectBestFeature-RFTree}
\alias{selectBestFeature}
\title{selectBestFeature-RFTree}
\usage{
selectBestFeature(x, y, featureList, sampleIndex = list(averagingSampleIndex =
  1:length(y), splittingSampleIndex = 1:length(y)),
  nodesize = list(splittingNodeSize = 5, averagingNodeSize = 5),
  splitrule = "variance")
}
\arguments{
\item{x}{A data frame or a matrix of all training predictors.}

\item{y}{A vector of all training responses.}

\item{featureList}{A list of candidate variables at the current split.}

\item{sampleIndex}{A list of index of dataset used in this node and its
children. `sampleIndex` contains two keys `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `splittingSampleIndex` is the same as `averagingSampleIndex`.}

\item{nodesize}{Minimum size of terminal nodes. Setting this number larger
causes smaller trees to be grown (and thus take less time). This parameter is
actually a list containing the values for both `splittingNodeSize` and
`averagingNodeSize`.}

\item{splitrule}{A splitting rule to determine the best split point among the
features. There are two possible split rules in the package, `variance` and
`maxstats`. The default is `variance` to minimize the overall MSE.}
}
\value{
A list of two outputs: "splitFeature" is the best feature to split
in order to minimize the split loss, "splitValue" is its corresponding split
value.
}
\description{
Find the best `splitfeature`, `splitValue` pair where
`splitfeature` is one of the features specified by `featureList`. The
`splitFeature` and its corresponding `splitValue` minimizes the specified
`splitrule`. The implementation is slightly different from the original
implementation as the tree contains both averaging and splitting dataset.
To check the minimum split, the method checks for both datasets according to
`sampleIndex` and `nodesize`.
}
