% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\docType{methods}
\name{showTree-RFTree}
\alias{showTree-RFTree}
\alias{showTree,RFTree-method}
\alias{showTree,}
\alias{RFTree-method}
\title{showTree-RFTree}
\usage{
showTree(object)

\S4method{showTree}{RFTree}(object)
}
\arguments{
\item{object}{RFTree object}
}
\description{
Print the entire tree starting from the root node
}
