% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExmpleSetups_causaleffects.R
\name{simulate_correlation_matrix}
\alias{simulate_correlation_matrix}
\title{Simulate a Correlation Matrix}
\usage{
simulate_correlation_matrix(dim, alpha)
}
\arguments{
\item{dim}{dimension of the correlation matrix.}

\item{alpha}{level of random correlation beteween the features. Choose 0 for
independent featues. The bigger alpha the bigger the correlation. For the
partial correlations the coefficients are created according to
beta(1/alpha, 1/alpha)}
}
\value{
A correlation matrix
}
\description{
This function uses the C-vine method for simulating correlation
  matrixes. Refer to the referenced paper for details.
}
\references{
\itemize{\item Daniel Lewandowskia, Dorota Kurowickaa, Harry Joe
  (2009). Generating random correlation matrices based on vines and extended
  onion method.}
}
