% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExmpleSetups_causaleffects.R
\docType{data}
\name{Lists Of Simulation Functions}
\alias{Lists Of Simulation Functions}
\alias{pscores.simulate_causal_experiment}
\alias{mu0.simulate_causal_experiment}
\alias{tau.simulate_causal_experiment}
\title{Lists Of Simulation Functions}
\format{An object of class \code{list} of length 5.}
\usage{
pscores.simulate_causal_experiment

mu0.simulate_causal_experiment

tau.simulate_causal_experiment
}
\description{
Lists Of Simulation Functions
}
\details{
These lists contain the example setups that can be used in
  \code{simulate_causal_experiment}. Run for example
  \code{tau.simulate_causal_experiment} to see the definition of the
  available functions
export
}
\keyword{datasets}
