% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRF.R
\docType{class}
\name{T_RF-class}
\alias{T_RF-class}
\title{This file implements the T-Learner (https://arxiv.org/pdf/1706.03461.pdf)
with the forestry implementation (https://github.com/soerenkuenzel/forestry)
as base learner.}
\description{
This file implements the T-Learner (https://arxiv.org/pdf/1706.03461.pdf)
with the forestry implementation (https://github.com/soerenkuenzel/forestry)
as base learner.
}
