% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thrf.R
\docType{data}
\name{T_RF-T_RF}
\alias{T_RF-T_RF}
\alias{T_RF}
\alias{T_RF,T_RF-T_RF}
\title{honstRF Constructor}
\format{An object of class \code{NULL} of length 0.}
\usage{
T_RF(feat, tr, yobs, mtry = ncol(feat), nodesizeSpl = 1, nodesizeAvg = 3,
  replace = TRUE, ntree = 1000, sample_fraction = 0.9, nthread = 4,
  splitratio = 0.5)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{mtry}{Number of variables to try at each node.}

\item{nodesizeSpl}{...}

\item{nodesizeAvg}{...}

\item{replace}{TODO: Add Description}

\item{ntree}{number of trees to grow}

\item{sample_fraction}{TODO: Add Description}

\item{nthread}{TODO: Add Description}

\item{splitratio}{TODO: Add Description}
}
\value{
A `T_RF` object.
}
\description{
This is an implementation of the T-learner combined with honest
random forest for both response functions

This is an implementation of the T-learner combined with honest
random forest for both response functions
}
\keyword{datasets}
