% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thrf.R
\docType{data}
\name{T_RF}
\alias{T_RF}
\title{honstRF Constructor}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{mtry}{Number of variables to try at each node.}

\item{nodesize}{...}

\item{replace}{...}

\item{ntree}{...}

\item{sample_fraction}{...}

\item{nthread}{...}

\item{splitratio}{...}

\item{nodesizeAvg}{...}
}
\description{
This is an implementation of the T-learner combined with honest
random forest for both response functions
}
\keyword{datasets}
