% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExmpleSetups_causaleffects.R
\docType{data}
\name{transphobia}
\alias{transphobia}
\title{Broockman and Kalla Transphobia data (cleaned)}
\format{
An object of class \code{data.frame} with 501 rows and 142 columns.
}
\usage{
transphobia
}
\description{
This is an example data set, and it has been created by looking at a
certain subset of the "Social Pressure and Voter Turnout: Evidence from a
Large-Scale Field Experiment" study that tested the impact of social pressure
on voter turnout. A precise description and the full data set can be found at
\url{https://isps.yale.edu/research/data/d001}.
}
\keyword{datasets}
