% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune_hyperband2.R
\name{tuneStageOne-honestRF}
\alias{tuneStageOne-honestRF}
\alias{tuneStageOne}
\title{tuneStageOne-honestRF}
\usage{
tuneStageOne(x, y, m_tau_init, sampsize = as.integer(nrow(x) * 0.75),
  num_iter = 1024, eta = 2, verbose = FALSE, seed = 24750371,
  nthread = 0)
}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{m_tau_init}{The initialize m_tau estimator.}

\item{sampsize}{The size of total samples to draw for the training data.}

\item{num_iter}{Maximum iterations/epochs per configuration. Default is 1024.}

\item{eta}{Downsampling rate. Default value is 2.}

\item{verbose}{if tuning process in verbose mode}

\item{seed}{random seed}

\item{nthread}{Number of threads to train and predict thre forest. The
default number is 0 which represents using all cores.}
}
\value{
A `honestRF` object
}
\description{
Tune the estimator in the first stage given a fixed second stage
estimator
}
