% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFNode.R
\name{weightedavgMean}
\alias{weightedavgMean}
\title{weightedavgMean}
\usage{
weightedavgMean(x, y, y_sd)
}
\arguments{
\item{x}{A dataframe of observations in the current tree leaf}

\item{y}{A vector of outcomes in the current tree leaf}

\item{y_sd}{A vector. The ith position contains the standard deviation of Y_i}
}
\value{
The weighted averaging of Y
}
\description{
A default function to average the observations in each tree leaf
by taking a weighted mean according with minimal variance.
}
