% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\name{X_BART}
\alias{X_BART}
\title{X_BART}
\usage{
X_BART(feat, tr, yobs, ensemble = "pscore", ndpost = 1200,
  tree_package = "dbarts", ntree = 200)
}
\arguments{
\item{feat}{A feature data frame.}

\item{tr}{A vector of treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{A vector of all the observed outcomes.}

\item{ensemble}{TODO: Add Description}

\item{ndpost}{TODO: Add Description}

\item{tree_package}{Package used to create tree.}

\item{ntree}{Number of trees to grow.}
}
\value{
A `X_BART` object.
}
\description{
This is an implementation of X_BART
}
\seealso{
Other metalearners: \code{\link{MO_BART}},
  \code{\link{S_BART}}, \code{\link{S_RF}},
  \code{\link{T_BART}}, \code{\link{T_RF}},
  \code{\link{X_RF}}
}
\concept{metalearners}
