% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRF.R
\docType{class}
\name{X_RF-class}
\alias{X_RF-class}
\title{This file implements the X-Learner (https://arxiv.org/pdf/1706.03461.pdf)
with the forestry implementation (https://github.com/soerenkuenzel/forestry)
as base learner.}
\description{
This file implements the X-Learner (https://arxiv.org/pdf/1706.03461.pdf)
with the forestry implementation (https://github.com/soerenkuenzel/forestry)
as base learner.
}
