% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\docType{data}
\name{X_RF-X_RF}
\alias{X_RF-X_RF}
\alias{X_RF}
\alias{X_RF,}
\alias{X_RF-X_RF}
\title{honstRF Constructor}
\format{An object of class \code{NULL} of length 0.}
\usage{
X_RF(feat, tr, yobs, predmode = "propmean", firststageVar = NULL,
  secondstageVar = NULL, num_trees_first = 50, num_trees_second = 50,
  mtry_first = ncol(feat)/2, mtry_second = ncol(feat),
  min_node_size_spl_first = 3, min_node_size_ave_first = 3,
  splitratio_first = 0.5, min_node_size_second = 1, replace_first = FALSE,
  replace_second = TRUE, sample_fraction_first = 0.7,
  sample_fraction_second = 0.5, prop_probability_tree = FALSE,
  honest_first = TRUE, honest_second = TRUE, honest_prop = TRUE,
  nthread = 4, min_node_size_spl_second = 5, splitratio_second = 0.5,
  min_node_size_ave_second = 5)
}
\arguments{
\item{feat}{A data frame of all training predictors.}

\item{tr}{A vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A vector containing the observed outcomes.}

\item{predmode}{default is "propmean".}

\item{firststageVar}{later}

\item{secondstageVar}{later}

\item{num_trees_first}{later}

\item{num_trees_second}{later}

\item{mtry_first}{later}

\item{mtry_second}{later}

\item{min_node_size_second}{later}

\item{replace_first}{later}

\item{replace_second}{later}

\item{sample_fraction_first}{later}

\item{sample_fraction_second}{later}

\item{prop_probability_tree}{later}

\item{honest_first}{later}

\item{honest_second}{later}

\item{honest_prop}{later}

\item{importance}{ok}

\item{min_node_size_first}{later}

\item{replace_first}{later}

\item{replace_second}{later}
}
\value{
A `X_RF` object.
}
\description{
Initialize a `honestRF` object.
}
\keyword{datasets}
