% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune_gpp.R
\name{X_RF_autotune_gpp}
\alias{X_RF_autotune_gpp}
\title{#' @title Autotuning for X-Learner with honest RF for both stages
#' @name X_RF_autotune_gpp
#' @rdname X_RF_autotune_gpp
#' @description This function tunes
#' @param feat A data frame of all the features.
#' @param tr A numeric vector contain 0 for control and 1 for treated variables.
#' @param yobs A numeric vector containing the observed outcomes.
#' @param ntree ..
#' @param Niter ..
#' @param nthread ..
#' @export X_RF_autotune_gpp}
\usage{
X_RF_autotune_gpp(feat, tr, yobs, ntree = 20000, nthread = 0,
  verbose = TRUE)
}
\description{
#' @title Autotuning for X-Learner with honest RF for both stages
#' @name X_RF_autotune_gpp
#' @rdname X_RF_autotune_gpp
#' @description This function tunes
#' @param feat A data frame of all the features.
#' @param tr A numeric vector contain 0 for control and 1 for treated variables.
#' @param yobs A numeric vector containing the observed outcomes.
#' @param ntree ..
#' @param Niter ..
#' @param nthread ..
#' @export X_RF_autotune_gpp
}
