% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRF_autotune_hyperband.R
\name{X_RF_autotune_hyperband}
\alias{X_RF_autotune_hyperband}
\title{Autotuning for X-Learner with honest RF for both stages}
\usage{
X_RF_autotune_hyperband(feat, tr, yobs, sample.fraction = 0.75,
  num_iter = 3^8, eta = 3, verbose = TRUE, seed = 24750371,
  nthread = 0)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{sample.fraction}{TODO: Add Description}

\item{num_iter}{Number of iterations.}

\item{eta}{Downsampling rate. Default value is 3.}

\item{verbose}{if tuning process in verbose mode}

\item{seed}{A random seed.}

\item{nthread}{Number of threads used to work in parallel.}
}
\description{
This function tunes
}
\examples{
  set.seed(14236142)
  feat <- iris[, -1]
  tr <- rbinom(nrow(iris), 1, .5)
  yobs <- iris[, 1]
  # train a
  xl_gpp <- X_RF_autotune_hyperband(feat, tr, yobs, verbose = FALSE)
  # computes the CATE and confidence intervals for CATE
  EstimateCate(xl_gpp, feat)
  CateCI(xl_gpp, feat, B = 5, verbose = FALSE)
}
\seealso{
\code{\link{X_RF_autotune_simple}}, \code{\link{X_RF_autotune_gpp}},
}
