% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune_simple.R
\name{X_RF_autotune_simple}
\alias{X_RF_autotune_simple}
\title{Autotuning for X-Learner with honest RF for both stages}
\usage{
X_RF_autotune_simple(feat, tr, yobs, ntree = 20000, nthread = 0,
  verbose = TRUE)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{ntree}{..}

\item{nthread}{..}

\item{Niter}{..}
}
\description{
This function tunes
}
