% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\docType{methods}
\name{CateCI-Meta-learner}
\alias{CateCI-Meta-learner}
\alias{CateCI,Meta-learner-method}
\title{CateCI-Meta-learner}
\usage{
\S4method{CateCI}{`Meta-learner`}(theObject, feature_new,
  method = "maintain_group_ratios", B = 200, nthread = 0,
  verbose = TRUE)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{feature_new}{A data frame.}

\item{method}{different versions of the bootstrap. Only n2TBS implemented}

\item{B}{number of bootstrap samples.}

\item{nthread}{number of threats used.}

\item{verbose}{...}
}
\value{
A data frame of estimated CATE Confidence Intervals
}
\description{
Return the estimated confidence intervals for the CATE
}
