% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{CateCI}
\alias{CateCI}
\title{Method CateCI}
\usage{
CateCI(theObject, feature_new, method = "maintain_group_ratios", B = 200,
  nthread = 0, verbose = TRUE)
}
\arguments{
\item{theObject}{...}

\item{feature_new}{...}

\item{method}{...}

\item{B}{number of bootstrap samples}

\item{nthread}{number of threads to be used in parallel}

\item{verbose}{...}
}
\description{
Method CateCI
}
