% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R, R/SBART.R, R/TBART.R,
%   R/XBART.R
\docType{methods}
\name{CateCI,CATE-estimator-method}
\alias{CateCI,CATE-estimator-method}
\alias{CateCI,Meta-learner-method}
\alias{CateCI,S_BART-method}
\alias{CateCI,T_BART-method}
\alias{CateCI,X_BART-method}
\title{CateCI-Meta-learner}
\usage{
\S4method{CateCI}{`CATE-estimator`}(theObject, feature_new,
  method = "maintain_group_ratios", bootstrapVersion = "normalApprox",
  B = 2000, nthread = 0, verbose = TRUE)

\S4method{CateCI}{S_BART}(theObject, feature_new, verbose = FALSE)

\S4method{CateCI}{T_BART}(theObject, feature_new, verbose = FALSE)

\S4method{CateCI}{X_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{theObject}{A `Meta-learner` object}

\item{feature_new}{A feature data frame}

\item{method}{Different versions of the bootstrap.}

\item{bootstrapVersion}{default is normalApprox which will just use the 
bootstrap normal approximation to get CI. smoothed will use use CI around the
smoothed bootstrap as introduced by Efron 2014.}

\item{B}{Number of bootstrap samples}

\item{nthread}{number of threads to be used in parallel}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A data frame of estimated CATE Confidence Intervals
}
\description{
Return the estimated confidence intervals for the CATE
}
