% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\docType{methods}
\name{EstimateAllSampleStatistics-Meta-learner}
\alias{EstimateAllSampleStatistics-Meta-learner}
\alias{EstimateAllSampleStatistics,Meta-learner-method}
\title{EstimateAllSampleStatistics-Meta-learner}
\usage{
\S4method{EstimateAllSampleStatistics}{`Meta-learner`}(theObject, method, B,
  nthread, verbose)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{method}{Different versions of the bootstrap.}

\item{B}{Number of boostrap samples.}

\item{nthread}{Number of threads to be used}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\description{
...
}
