% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\docType{methods}
\name{EstimateAllSampleStatistics-Meta-learner}
\alias{EstimateAllSampleStatistics-Meta-learner}
\title{EstimateAllSampleStatistics-Meta-learner}
\usage{
\S4method{EstimateAllSampleStatistics}{`Meta-learner`}(theObject, method, B,
  nthread, verbose)
}
\description{
Return the estimated CATE
}
