% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBART.R
\docType{methods}
\name{EstimateAllSampleStatistics-S_BART}
\alias{EstimateAllSampleStatistics-S_BART}
\title{EstimateAllSampleStatistics-S_BART}
\usage{
\S4method{EstimateAllSampleStatistics}{S_BART}(theObject, verbose = FALSE)
}
\description{
Return the estimated CATE
}
