% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\docType{methods}
\name{EstimateAllSampleStatistics-X_BART}
\alias{EstimateAllSampleStatistics-X_BART}
\title{EstimateAllSampleStatistics-X_BART}
\usage{
\S4method{EstimateAllSampleStatistics}{X_BART}(theObject, verbose = FALSE)
}
\description{
Return the estimated CATE
}
