% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRF.R
\docType{methods}
\name{EstimateAllSampleStatistics-X_RF}
\alias{EstimateAllSampleStatistics-X_RF}
\alias{EstimateAllSampleStatistics,X_RF-method}
\title{EstimateAllSampleStatistics-X_RF}
\usage{
\S4method{EstimateAllSampleStatistics}{X_RF}(theObject, method, B, nthread,
  verbose)
}
\arguments{
\item{theObject}{A `X_hRF` object.}

\item{method}{different versions of the bootstrap.}

\item{B}{Number of bootstrap samples.}

\item{nthread}{Number of threads to be used.}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\description{
Return the estimated CATE
}
