% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\docType{methods}
\name{EstimateAllSampleStatistics-X_RF}
\alias{EstimateAllSampleStatistics-X_RF}
\title{EstimateAllSampleStatistics-X_RF}
\usage{
\S4method{EstimateAllSampleStatistics}{X_RF}(theObject, method, B, nthread,
  verbose)
}
\description{
Return the estimated CATE
}
