% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R, R/SBART.R, R/TBART.R,
%   R/XBART.R
\docType{methods}
\name{EstimateAllSampleStatistics}
\alias{EstimateAllSampleStatistics}
\alias{EstimateAllSampleStatistics,CATE-estimator-method}
\alias{EstimateAllSampleStatistics,Meta-learner-method}
\alias{EstimateAllSampleStatistics,S_BART-method}
\alias{EstimateAllSampleStatistics,T_BART-method}
\alias{EstimateAllSampleStatistics,X_BART-method}
\title{Method EstimateAllSampleStatistics}
\usage{
EstimateAllSampleStatistics(theObject, feature_new,
  method = "maintain_group_ratios", B = 2000, nthread = 0,
  verbose = TRUE)

\S4method{EstimateAllSampleStatistics}{`CATE-estimator`}(theObject, method,
  B, nthread, verbose)

\S4method{EstimateAllSampleStatistics}{S_BART}(theObject,
  verbose = FALSE)

\S4method{EstimateAllSampleStatistics}{T_BART}(theObject,
  verbose = FALSE)

\S4method{EstimateAllSampleStatistics}{X_BART}(theObject,
  verbose = FALSE)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{feature_new}{A feature data frame}

\item{method}{different versions of the bootstrap.}

\item{B}{Number of bootstrap samples.}

\item{nthread}{Number of threads to be used.}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\description{
TODO: add description
}
