% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{EstimateATT}
\alias{EstimateATT}
\title{Method EstimateATT}
\usage{
EstimateATT(theObject, feature_new, method = "maintain_group_ratios",
  B = 2000, nthread = 0, verbose = TRUE)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}

\item{method}{Different versions of the bootstrap.}

\item{B}{Number of bootstrap samples.}

\item{nthread}{Number of threads used in parallel.}

\item{verbose}{TRUE for detailed output, FALSE for no output.}
}
\description{
Method EstimateATT
}
