% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBART.R
\name{EstimateCate,S_BART-method}
\alias{EstimateCate,S_BART-method}
\title{EstimateCate-S_BART}
\usage{
\S4method{EstimateCate}{S_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}
}
\description{
Returns the estimated CATE.
}
