% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRF.R, R/SRF.R
\name{EstimateCate,M_RF-method}
\alias{EstimateCate,M_RF-method}
\alias{EstimateCate,S_RF-method}
\title{EstimateCate-S_hRF}
\usage{
\S4method{EstimateCate}{M_RF}(theObject, feature_new)

\S4method{EstimateCate}{S_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}
}
\description{
Returns the estimated CATE.
}
