% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBART.R
\docType{methods}
\name{EstimateCate-T_BART}
\alias{EstimateCate-T_BART}
\alias{EstimateCate,}
\alias{T_BART-method}
\alias{EstimateCate-T_BART}
\alias{EstimateCate,}
\alias{T_BART-method}
\title{EstimateCate-T_BART}
\usage{
\S4method{EstimateCate}{T_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{CateCI}{T_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{feature_new}{A data frame.}

\item{verbose}{Should the training output be posted?}

\item{object}{A `T_BART` object.}

\item{object}{A `T_BART` object.}

\item{feature_new}{A data frame.}

\item{verbose}{Should the training output be posted?}
}
\value{
A vector of predicted CATE

A vector of predicted CATE
}
\description{
Return the estimated CATE

Return the estimated CATE
}
