% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRF.R
\name{EstimateCate,T_RF-method}
\alias{EstimateCate,T_RF-method}
\title{EstimateCate-T_hRF}
\usage{
\S4method{EstimateCate}{T_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}
}
\description{
Returns the estimated CATE.
}
