% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\name{EstimateCate,X_BART-method}
\alias{EstimateCate,X_BART-method}
\title{EstimateCate-X_BART}
\usage{
\S4method{EstimateCate}{X_BART}(theObject, feature_new, verbose = FALSE, return_CI = FALSE)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}
}
\description{
Returns the estimated CATE.
}
