% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\docType{methods}
\name{EstimateCate-X_BART}
\alias{EstimateCate-X_BART}
\alias{EstimateCate,}
\alias{X_BART-method}
\alias{EstimateCate-X_BART}
\alias{EstimateCate,}
\alias{X_BART-method}
\title{EstimateCate-X_BART}
\usage{
\S4method{EstimateCate}{X_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{CateCI}{X_BART}(theObject, feature_new, verbose = FALSE)
}
\arguments{
\item{feature_new}{A data frame.}

\item{verbose}{Should the training output be posted?}

\item{object}{A `X_BART` object.}

\item{ensemble}{Specify how tau_hat_1 and tau_hat_2 should be averaged the
dafault is by the weight of the propensity score, other version is
"variance" which uses the variance of the weighting}

\item{object}{A `X_BART` object.}

\item{feature_new}{A data frame.}

\item{verbose}{Should the training output be posted?}
}
\value{
A vector of predicted CATE

A vector of predicted CATE
}
\description{
Return the estimated CATE

Return the estimated CATE
}
