% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRF.R
\name{EstimateCate,X_RF-method}
\alias{EstimateCate,X_RF-method}
\title{EstimateCate-X_hRF}
\usage{
\S4method{EstimateCate}{X_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}
}
\description{
Returns the estimated CATE.
}
