% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{EstimateCate}
\alias{EstimateCate}
\title{Method EstimateCate}
\usage{
EstimateCate(theObject, feature_new, ...)
}
\arguments{
\item{theObject}{A `MetaLearner` object.}

\item{feature_new}{A feature data frame.}

\item{...}{Additional parameters that are specific for some MetaLearners}
}
\description{
Returns the estimated CATE.
}
