% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R, R/MRF.R, R/SRF.R, R/SBART.R,
%   R/XRF.R, R/XBART.R, R/TRF.R, R/TBART.R, R/MBART.R
\docType{methods}
\name{EstimateCate}
\alias{EstimateCate}
\alias{EstimateCate,M_RF-method}
\alias{EstimateCate,S_RF-method}
\alias{EstimateCate,S_BART-method}
\alias{EstimateCate,X_RF-method}
\alias{EstimateCate,X_BART-method}
\alias{EstimateCate,T_RF-method}
\alias{EstimateCate,T_BART-method}
\alias{EstimateCate,M_BART-method}
\title{Method EstimateCate}
\usage{
EstimateCate(theObject, feature_new, ...)

\S4method{EstimateCate}{M_RF}(theObject, feature_new)

\S4method{EstimateCate}{S_RF}(theObject, feature_new)

\S4method{EstimateCate}{S_BART}(theObject, feature_new, verbose = FALSE)

\S4method{EstimateCate}{X_RF}(theObject, feature_new)

\S4method{EstimateCate}{X_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{EstimateCate}{T_RF}(theObject, feature_new)

\S4method{EstimateCate}{T_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{EstimateCate}{M_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{feature_new}{A feature data frame.}

\item{...}{Additional parameters that are specific for some meta-learners}

\item{verbose}{TRUE for detailed output FALSE for no output}

\item{return_CI}{If TRUE, return predictions and their confidence intervals;}
}
\description{
Returns the estimated CATE.
}
