% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R, R/MOBART.R, R/SBART.R,
%   R/SRF.R, R/TBART.R, R/XBART.R, R/XRF.R, R/TRF.R
\docType{methods}
\name{EstimateCate}
\alias{EstimateCate}
\alias{EstimateCate-M_BART}
\alias{EstimateCate-S_BART}
\alias{EstimateCate-S_RF}
\alias{EstimateCate-X_BART}
\alias{EstimateCate-X_RF}
\alias{EstimateCate-T_RF}
\title{Method EstimateCate}
\usage{
EstimateCate(theObject, feature_new, ...)

\S4method{EstimateCate}{MO_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{EstimateCate}{S_BART}(theObject, feature_new, verbose = FALSE)

\S4method{EstimateCate}{S_RF}(theObject, feature_new)

\S4method{EstimateCate}{T_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{EstimateCate}{X_BART}(theObject, feature_new, verbose = FALSE,
  return_CI = FALSE)

\S4method{EstimateCate}{X_RF}(theObject, feature_new)

\S4method{EstimateCate}{T_RF}(theObject, feature_new)
}
\arguments{
\item{theObject}{A `Meta-learner` object.}

\item{feature_new}{A feature data frame.}

\item{...}{Additional parameters that are specific for some meta-learners}

\item{verbose}{TRUE for detailed output FALSE for no output}

\item{return_CI}{If TRUE, return predictions and their confidence intervals;}

\item{theObject}{A `X_BART` object.}

\item{feature_new}{A data frame.}

\item{verbose}{TRUE for detailed output FALSE for no output}

\item{return_CI}{TRUE for predictions and confidence intervals FALSE for only}
}
\description{
Returns the estimated CATE.
}
