% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE_estimators.R
\name{EstimateCate}
\alias{EstimateCate}
\title{Method EstimateCate}
\usage{
EstimateCate(theObject, feature_new, ...)
}
\arguments{
\item{theObject}{...}

\item{feature_new}{...}

\item{...}{Additional parameters}
}
\description{
Method EstimateCate
}
