% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\name{getOOB-honestRF}
\alias{getOOB-honestRF}
\title{getOOB-honestRF}
\usage{
getOOB(object, noWarning = FALSE)
}
\arguments{
\item{object}{A `honestRF` object.}

\item{noWarning}{flag to not display warnings}
}
\description{
Calculate the out-of-bag error of a given forest.
}
