% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_double_robust.R
\name{gof_double_robust}
\alias{gof_double_robust}
\title{gof_double_robust}
\usage{
gof_double_robust(feat, yobs, tr, estimator, emin, k = 5, verbose = FALSE)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observed values}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{learner constructor}

\item{emin}{the pscore estimate will be bounded between emin and 1 - emin
in order to avoid any divide-by-zero errors}

\item{k}{number of folds used for cross-validation}

\item{verbose}{logical. If TRUE then detailed updates will be printed}
}
\value{
mean(error) and sd(error)
}
