% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_matching.R
\name{gof_matching}
\alias{gof_matching}
\title{gof_matching}
\usage{
gof_matching(feat, yobs, tr, estimator, estimand = "ATT", k = 2,
  replace = TRUE, emin = 1e-05)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observations}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{estimand}{ATE, ATT or ATC}

\item{k}{we are doing a k fold cross validation}
}
\value{
mean(error) and sd(error)
}
