% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_matching.R
\name{gof_matching}
\alias{gof_matching}
\title{gof_matching}
\usage{
gof_matching(feat, yobs, tr, estimator, estimand = "ATT", k = 2,
  replace = TRUE, emin = 1e-05, verbose = FALSE)
}
\arguments{
\item{feat}{a data frame of features}

\item{yobs}{a vector of observed values}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{learner constructor}

\item{estimand}{ATE, ATT or ATC}

\item{k}{number of folds used for cross-validation}

\item{replace}{logical. If TRUE then samples will be replaced while matching}

\item{emin}{the pscore estimate will be bounded between emin and 1 - emin
in order to avoid any divide-by-zero errors}

\item{verbose}{logical. If TRUE then detailed updates will be printed}
}
\value{
mean(error) and sd(error)
}
\description{
Match each unit in one group(treatment/control) to units in
another group(treatment/control) based on the chosen estimand. For each iteration
of cross validation, assign treatment lablels (0 or 1) with probability given 
by averaged propensity score for each pair of match and estimate CATE. Return
the mean squared error and its standard deviation of CATE from the true difference.
}
