% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_transformed.R
\docType{methods}
\name{gof_transformed}
\alias{gof_transformed}
\title{gof_transformed}
\usage{
\S4method{gof_transformed}{ANY}(feat, yobs, tr, estimator, k = 2,
  emin = 1e-05)
}
\arguments{
\item{feat}{a data frame of features}

\item{tr}{a vector of group assignment (assume entries are integers)}

\item{estimator}{a learner constructor}

\item{k}{we are doing a k fold cross validation}

\item{emin}{the pscore prediciton will be bounded between emin and 1- emin
to aovid decide by 0 error}

\item{obs}{a vector of observations}
}
\description{
gof_transformed
}
