% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector_transformed.R
\name{gof_transformed}
\alias{gof_transformed}
\title{gof_transformed}
\usage{
gof_transformed(feat, yobs, tr, estimator, k = 2, emin = 1e-05,
  verbose = FALSE)
}
\arguments{
\item{feat}{A data frame of features}

\item{yobs}{A vector of observations}

\item{tr}{A vector of group assignment (assume entries are integers)}

\item{estimator}{A learner constructor}

\item{k}{Number of folds used for cross validation}

\item{emin}{the pscore prediciton will be bounded between emin and 1 - emin
to avoid decide by 0 error}

\item{verbose}{determines whether detailed updates will be printed}
}
\value{
mean(error) and sd(error)
}
\description{
gof_transformed
}
