% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExmpleSetups_causaleffects.R
\docType{data}
\name{gotv}
\alias{gotv}
\title{Get Out To Vote}
\format{An object of class \code{data.frame} with 229461 rows and 9 columns.}
\usage{
gotv
}
\description{
This is an example data set, and it has been created by looking at a
certain subset of the "Social Pressure and Voter Turnout: Evidence from a
Large-Scale Field Experiment" study that tested the impact of social pressure
on voter turnout. A precise description and the full data set can be found at
\url{https://isps.yale.edu/research/data/d001}. 

The study consists of seven key, individual-level covariates, most of which
are discrete: gender, age, and whether the registered individual voted in the
primary elections in 2000, 2002, and 2004 or the general elections in 2000 and
2002. The sample was restricted to voters who had voted in the 2004 general
election. The outcome of interest was the turnout in the 2006 primary election,
which was an indicator variable, and the treatment was whether or not the
subjects were elected to receive a mailer.
}
\references{
\itemize{
  \item Gerber AS, Green DP, Larimer CW (2008) Social Pressure and Voter
  Turnout: Evidence from a Large-Scale Field Experiment. Am Polit Sci Rev
  102:33–48. \url{https://isps.yale.edu/research/publications/isps08-001}
 }
}
\keyword{datasets}
