% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRFTree.R
\docType{class}
\name{honestRFTree-class}
\alias{honestRFTree-class}
\alias{honestRFTree}
\title{honestRFTree Constructor}
\usage{
honestRFTree(x, y, mtry = max(floor(ncol(x)/3), 1),
  nodesize = list(averagingNodeSize = 5, splittingNodeSize = 5),
  sampleIndex = list(averagingSampleIndex = 1:length(y), splittingSampleIndex
  = 1:length(y)), splitrule = "variance")
}
\arguments{
\item{x}{A data frame or a matrix of all training predictors.}

\item{y}{A vector of all training responses.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
The default value is set to be one third of total feature amount.}

\item{nodesize}{Minimum size of terminal nodes. Setting this number larger
causes smaller trees to be grown (and thus take less time). For the
`honestRFTree`, the minimum size of nodes need to be specified for both
averaging dataset and splitting dataset. The default value for both is 5.}

\item{sampleIndex}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset. For the `honestRFTree`, the sampleIndex need to be specified for
both averaging dataset and splitting dataset. The default value for both is
all the dataset of input dataset.}

\item{splitrule}{A splitting rule to determine the best split point among the
features. There are two possible split rules in the package, `variance` and
`maxstats`. The default is `variance` to minimize the overall MSE.}
}
\value{
a `honestRFTree` object
}
\description{
`honestRFTree` inherits `RFTree`, which serves as a modified
version of `RFTree`. The major change is that when `honestRFTree` determines
a valid split point, it looks at whether or not it is valid for both
averaging and splitting dataset.

honestRFTree Constructor
}
