% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\docType{methods}
\name{predict-honestRF}
\alias{predict-honestRF}
\alias{predict,}
\alias{honestRF-method}
\title{predict-honestRF}
\usage{
\S4method{predict}{honestRF}(object, feature.new)
}
\arguments{
\item{object}{A `honestRF` object.}

\item{feature.new}{A data frame of testing predictors.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the forest.
}
