% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\docType{methods}
\name{predict-RF}
\alias{predict-RF}
\alias{predict,}
\alias{RF-method}
\title{predict-RF}
\usage{
\S4method{predict}{RF}(object, feature.new)
}
\arguments{
\item{object}{RF object.}

\item{feature.new}{A data frame or a matrix of all testing predictors.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the current forest.
}
