% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFNode.R
\docType{methods}
\name{predict-RFNode}
\alias{predict-RFNode}
\alias{predict,}
\alias{RFNode-method}
\title{predict-RFNode}
\usage{
\S4method{predict}{RFNode}(object, feature.new, x, y, se = NULL, avgfunc,
  categoricalFeatureCols)
}
\arguments{
\item{object}{A `RFNode`` object.}

\item{feature.new}{A data frame of testing predictors.}

\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{avgfunc}{An averaging function to average observations in the node. The
function is used for prediction. The input of this function should be a
dataframe of predictors `x` and a vector of outcomes `y`. The output is a
scalar. The default function is to take the mean of vector `y`.}

\item{categoricalFeatureCols}{A list of index for all categorical data. Used
for trees to detect categorical columns.}
}
\value{
A vector of predicted responses.
}
\description{
Return the prediction from the current node if it is a leaf
node, otherwise it will recursively call its children according to the
`splitFeature` and `splitValue`.
}
