% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\name{recursivePartition-RFTree}
\alias{recursivePartition-RFTree}
\alias{recursivePartition}
\title{recursivePartition-RFTree}
\usage{
recursivePartition(x, y, mtry = max(floor(ncol(x)/3), 1),
  sampleIndex = list(averagingSampleIndex = 1:length(y), splittingSampleIndex
  = 1:length(y)), nodesize = list(splittingNodeSize = 5, averagingNodeSize =
  5), splitrule = "variance", categoricalFeatureCols = list())
}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{mtry}{The number of variables randomly selected at each split point.
The default value is set to be one third of total number of features of the
training data.}

\item{sampleIndex}{A list of index of dataset used in this node and its
children. `sampleIndex` contains two keys `averagingSampleIndex`
and `splittingSampleIndex`. `averagingSampleIndex` is used to generate
aggregated prediction for the node. `splittingSampleIndex` is used for
`honestRF` which stores the splitting data when creating the tree. In
default, `splittingSampleIndex` is the same as `averagingSampleIndex`.}

\item{nodesize}{The minimum observations contained in terminal nodes. This
parameter is actually a list containing the values for both
`splittingNodeSize` and `averagingNodeSize`.}

\item{splitrule}{A string to specify how to find the best split among all
candidate feature values. The current version only supports `variance` which
minimizes the overall MSE after splitting. The default value is `variance`.}

\item{categoricalFeatureCols}{A list of index for all categorical data. Used
for trees to detect categorical columns.}
}
\value{
root node.
}
\description{
Grow the decision tree by recursively finding the best split
feature and value.
}
