% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\docType{class}
\name{RF-class}
\alias{RF-class}
\title{RF-Constructor}
\description{
`RF` object implementing the most basic version of a random
forest.
}
\section{Slots}{

\describe{
\item{\code{x}}{A data frame of all training predictors.}

\item{\code{y}}{A vector of all training responses.}

\item{\code{ntree}}{The number of trees to grow in the forest. The default value is
500.}

\item{\code{replace}}{An indicator of whether sampling of training data is with
replacement. The default value is TRUE.}

\item{\code{sampsize}}{The size of total samples to draw for the training data. If
sampling with replacement, the default value is the length of the training
data. If samplying without replacement, the default value is two-third of
the length of the training data.}

\item{\code{mtry}}{The number of variables randomly selected at each split point.
The default value is set to be one third of total number of features of the
training data.}

\item{\code{nodesize}}{The minimum observations contained in terminal nodes. The
default value is 5.}

\item{\code{splitrule}}{A string to specify how to find the best split among all
candidate feature values. The current version only supports `variance` which
minimizes the overall MSE after splitting. The default value is `variance`.}

\item{\code{avgfunc}}{An averaging function to average observations in the node. The
function is used for prediction. The input of this function should be a
dataframe of predictors `x` and a vector of outcomes `y`. The output is a
scalar. The default function is to take the mean of vector `y`.}

\item{\code{forest}}{A list of `RFTree` objects in the forest. If the class is
extended, the list may contain the corresponding extended `RFTree` object.}

\item{\code{encodingLabels}}{A list contains all encodings that are used to encode
categorical columns in the training data.}
}}

