% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFTree.R
\docType{class}
\name{RFTree-class}
\alias{RFTree-class}
\alias{RFTree}
\title{RFTree Constructor}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{y}{A vector of all training responses.}

\item{mtry}{The number of variables randomly selected at each split point.
The default value is set to be one third of total number of features of the
training data.}

\item{nodesize}{The minimum observations contained in terminal nodes. The
default value is 5.}

\item{sampleIndex}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}

\item{splitrule}{A string to specify how to find the best split among all
candidate feature values. The current version only supports `variance` which
minimizes the overall MSE after splitting. The default value is `variance`.}

\item{categoricalFeatureCols}{A list of index for all categorical data. Used
for trees to detect categorical columns.}
}
\description{
`RFTree` is the unit component in the `RF` which composes
`RFNode`. The tree uses recursively partitioning to determine the best
`splitFeature` and `splitValue` for each level, and recursively split the
dataset until it reaches the limitation according to `nodesize`.

The actual storation of `RFTree` is different from its input.
`sampleIndex` and `nodesize` are actually two lists composing of those in
splitting dataset and averaging dataset. In the original random forest,
those are the same value. For other variant of random forest, such as honest
random forest, they can be set differently. The constructor can be
overwritten to reflect the change.
}
\section{Slots}{

\describe{
\item{\code{sampleIndex}}{A list of the index of observations that are used as
averaging dataset. The index are based on the original dataset `x` and `y`
from forest. Essentially, `x[sampleIndex]` generates the whole splitting
dataset.}

\item{\code{root}}{A `RFNode` object which is the root of the tree. If the class is
extended, the list may contain the corresponding extended `RFNode` object.}
}}

\keyword{datasets}
