% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBART.R
\name{S_BART}
\alias{S_BART}
\title{S_BART}
\usage{
S_BART(feat, tr, yobs, ndpost = 1200,
  sample_stat = "counterfactuals estimated", tree_package = "dbarts",
  ntree = 200, verbose)
}
\arguments{
\item{feat}{A data frame of features}

\item{tr}{A vector of treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{A vector of observed outcomes.}

\item{ndpost}{TODO: Add description}

\item{sample_stat}{TODO: Add description}

\item{tree_package}{Package used to create a tree}

\item{ntree}{Number of trees to grow}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A `S_BART` object.
}
\description{
This is an implementation of S_BART
}
