% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrf.R
\docType{class}
\name{S_RF-class}
\alias{S_RF-class}
\title{ShRF constructor}
\description{
The `S_RF` object is S-learner combined with honest random
forest used for both response functions
}
\section{Slots}{

\describe{
\item{\code{feature_train}}{A data frame of all training features.}

\item{\code{tr_train}}{A vector contain 0 for control and 1 for treated variables.}

\item{\code{yobs_train}}{A vector containing the observed outcomes.}

\item{\code{m_y_t}}{contains an honest random forest predictor for the treated group}

\item{\code{m_y_c}}{contains an honest random forest predictor for the control group}

\item{\code{creator}}{function which creates a S_RF}
}}

