% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrf.R
\name{S_RF}
\alias{S_RF}
\alias{S_RF,S_RF-S_RF}
\title{S_RF Constructor}
\usage{
S_RF(feat, tr, yobs, mtry = ncol(feat), nodesizeSpl = 1, nodesizeAvg = 3,
  replace = TRUE, ntree = 1000, sample_fraction = 0.9, nthread = 4,
  splitratio = 0.5, alwaysTr = FALSE)
}
\value{
A `S_RF` object.
}
\description{
This is an implementation of the T-learner combined with honest
  random forest for both response functions
}
