% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrf.R
\docType{data}
\name{S_RF-S_RF}
\alias{S_RF-S_RF}
\alias{S_RF}
\alias{S_RF,S_RF-S_RF}
\title{honstRF Constructor}
\format{An object of class \code{NULL} of length 0.}
\usage{
S_RF(feat, tr, yobs, mtry = ncol(feat), nodesizeSpl = 1, nodesizeAvg = 3,
  replace = TRUE, ntree = 1000, sample_fraction = 0.9, nthread = 4,
  splitratio = 0.5, alwaysTr = FALSE)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{mtry}{Number of variables to try at each node.}

\item{nodesizeSpl}{...}

\item{nodesizeAvg}{TODO: Add Description}

\item{replace}{TODO: Add Description}

\item{ntree}{TODO: Add Description}

\item{sample_fraction}{TODO: Add Description}

\item{nthread}{TODO: Add Description}

\item{splitratio}{TODO: Add Description}

\item{alwaysTr}{weather or not we always test weather we should split on the
treatment assignment. Currently only alwaysTr=FALSE is implemented}
}
\value{
A `S_RF` object.
}
\description{
This is an implementation of the S-learner combined with honest
  random forest for both response functions

This is an implementation of the S-learner combined with honest
  random forest for both response functions
}
\keyword{datasets}
