% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExmpleSetups_causaleffects.R
\name{simulate_causal_experiment}
\alias{simulate_causal_experiment}
\title{simulate a RCT or observational data for causal effect estimation}
\usage{
simulate_causal_experiment(ntrain, ntest, dim, alpha = 0.1,
  feat_distribution = "normal", given_features = NULL,
  setup = "RespSparseTau1strong")
}
\arguments{
\item{ntrain}{Number of training examples.}

\item{ntest}{Number of test examples.}

\item{dim}{Dimension of the data set.}

\item{alpha}{Only used if given_features is not set and feat_distribution is
chosen to be normal. It specifies how correlated the features can be. If
alpha = 0, then the features are independent if alpha is very large, then the
features can be very correlated.}

\item{feat_distribution}{Only used if given_features is not specified. Either
"normal" or "unif". It specifies the distribution of the features.}

\item{given_features}{This is used, if we already have features and want to test
the performance of different estimators for a particular set of features.}

\item{setup}{This is used to specify the function form of the potential
outcomes and the treatment assignment.}
}
\value{
A list of the transformed object `x`, and encoding information
`labels`.
}
\description{
This function simulates a RCT or observational data for causal
effect estimation to test different heterogenuous treatment effect estimation
strategies.
}
