% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExmpleSetups_causaleffects.R
\name{simulate_correlation_matrix}
\alias{simulate_correlation_matrix}
\title{simulate a RCT or observational data for causal effect estimation}
\usage{
simulate_correlation_matrix(dim, alpha)
}
\arguments{
\item{dim}{Number of training examples.}

\item{alpha}{level of random correlation beteween the features. Choose 0 for
independent featues. The bigger alpha the bigger the correlation. For the
partial correlations the coefficients are created according to
beta(1/alpha, 1/alpha)}
}
\value{
A correlation matrix
}
\description{
This function uses the C-vine method for simulating correlation
  matrixes as in the paper "Generating random correlation matrices based on
  vines and extended onion method" by Daniel Lewandowskia, Dorota Kurowickaa,
  Harry Joe
}

