% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBART.R
\name{T_BART}
\alias{T_BART}
\title{T_BART}
\usage{
T_BART(feat, tr, yobs, ndpost = 1200,
  sample_stat = "counterfactuals estimated", tree_package = "dbarts",
  ntree = 200, verbose = FALSE)
}
\arguments{
\item{feat}{A feature data.frame.}

\item{tr}{A vector of treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{A vector of the observed outcome.}

\item{ndpost}{...}

\item{sample_stat}{...}

\item{tree_package}{Name of the package used to create a tree}

\item{ntree}{Number of trees to grow}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A `T_BART` object.
}
\description{
This is an implementation of T_BART
}
