% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBART.R
\name{T_BART}
\alias{T_BART}
\title{T_BART}
\usage{
T_BART(feat, tr, yobs, ndpost = 1200,
  sample_stat = "counterfactuals estimated", verbose)
}
\arguments{
\item{feat}{feature data.frame.}

\item{tr}{treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{the observed outcome.}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A `X_RF` object.
}
\description{
This is an implementation of T_BART
}
