% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thrf.R
\name{T_RF}
\alias{T_RF}
\alias{T_RF,}
\alias{T_RF-T_RF}
\title{T_RF Constructor}
\usage{
T_RF(feat, tr, yobs, mtry = ncol(feat), nodesizeSpl = 1, nodesizeAvg = 3,
  replace = TRUE, ntree = 500, sample_fraction = 0.9, nthread = 4,
  splitratio = 0.5)
}
\value{
A `T_RF` object.
}
