% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honestRF.R
\name{testing_data_checker-honestRF}
\alias{testing_data_checker-honestRF}
\alias{testing_data_checker}
\title{testing_data_checker-hoenstRF}
\usage{
testing_data_checker(x, feature.new)
}
\arguments{
\item{x}{A data frame of all training predictors.}

\item{feature.new}{A data frame of testing predictors.}
}
\description{
Check the testing data to do prediction
}
