% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\name{X_BART}
\alias{X_BART}
\title{X_BART}
\usage{
X_BART(feat, tr, yobs, ensemble = "pscore", ndpost = 1200,
  tree_package = "dbarts", ntree = 200)
}
\arguments{
\item{feat}{A feature dataframe.}

\item{tr}{A vector of treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{A vector the observed outcome.}

\item{ensemble}{...}

\item{ndpost}{...}

\item{tree_package}{Package used to create tree}

\item{ntree}{Number of trees to grow}
}
\value{
A `X_RF` object.
}
\description{
This is an implementation of X_BART
}
