% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBART.R
\name{X_BART}
\alias{X_BART}
\title{X_BART}
\usage{
X_BART(feat, tr, yobs, ensemble = "pscore", ndpost = 1200)
}
\arguments{
\item{feat}{feature data.frame.}

\item{tr}{treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{the observed outcome.}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A `X_RF` object.
}
\description{
This is an implementation of X_BART
}
