% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\docType{class}
\name{X_RF-class}
\alias{X_RF-class}
\title{XhRF constructor}
\description{
The `X_RF` object is X-learner combined with honest random
forest used for the propensity score estimate, the first stage and the second
stage.
}
\section{Slots}{

\describe{
\item{\code{feature_train}}{A data frame of all training features.}

\item{\code{tr_train}}{A vector contain 0 for control and 1 for treated variables.}

\item{\code{yobs_train}}{A vector containing the observed outcomes.}

\item{\code{m_0}}{contains an honest random forest predictor for the control group of
the first stage.}

\item{\code{m_1}}{contains an honest random forest predictor for the treated group of
the first stage.}

\item{\code{m_tau_0}}{contains an honest random forest predictor for the control
group of the second stage.}

\item{\code{m_tau_1}}{contains an honest random forest predictor for the treated
group of the second stage.}

\item{\code{m_prop}}{contains an honest random forest predictor the propensity score.}

\item{\code{firststageVar}}{contains the indices of variables, which are only used in
the first stage.}

\item{\code{secondstageVar}}{contains the numbers of variables, which are only used
in the second stage.}
}}

