% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRF.R
\name{X_RF-X_RF}
\alias{X_RF-X_RF}
\alias{X_RF}
\title{X-Learner with honest RF for both stages}
\usage{
X_RF(feat, tr, yobs, predmode = "propmean",
  relevant_Variable_first = 1:ncol(feat),
  relevant_Variable_second = 1:ncol(feat),
  relevant_Variable_prop = 1:ncol(feat), ntree_first = 1000,
  ntree_second = 1000, ntree_prop = 500,
  mtry_first = round(ncol(feat) * 13/20),
  mtry_second = round(ncol(feat) * 17/20), mtry_prop = ncol(feat),
  min_node_size_spl_first = 2, min_node_size_ave_first = 1,
  min_node_size_spl_second = 5, min_node_size_ave_second = 6,
  min_node_size_spl_prop = 11, min_node_size_ave_prop = 33,
  splitratio_first = 1, splitratio_second = 0.8,
  splitratio_prop = 0.5, replace_first = TRUE, replace_second = TRUE,
  replace_prop = TRUE, sample_fraction_first = 0.8,
  sample_fraction_second = 0.7, sample_fraction_prop = 0.5,
  nthread = 0, verbose = TRUE, middleSplit_first = TRUE,
  middleSplit_second = TRUE, middleSplit_prop = FALSE)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector containing 0 for control and 1 for treated 
variables.}

\item{yobs}{A numeric vector containing all the observed outcomes.}

\item{predmode}{One of propmean, control, treated, extreme. It specifies how
the two estimators of the second stage should be aggregated. The default is
propmean which refers to propensity score weighting.}

\item{relevant_Variable_first}{Variables which are only used 
in the first stage.}

\item{relevant_Variable_second}{Variables which are only used 
in the second stage.}

\item{relevant_Variable_prop}{TODO: Add Description}

\item{ntree_first}{Numbers of trees in the first stage.}

\item{ntree_second}{Numbers of trees in the second stage.}

\item{ntree_prop}{TODO: Add Description}

\item{mtry_first}{Numbers of trees in the first stage.}

\item{mtry_second}{Numbers of trees in the second stage.}

\item{mtry_prop}{TODO: Add Description}

\item{min_node_size_spl_first}{minimum nodesize in the first stage for the
observations in the splitting set.}

\item{min_node_size_ave_first}{minimum nodesize in the first stage for the
observations in the averaging set.}

\item{min_node_size_spl_second}{minimum nodesize in the second stage for the
observations in the splitting set.}

\item{min_node_size_ave_second}{minimum nodesize in the second stage for the
observations in the averaging set.}

\item{min_node_size_spl_prop}{TODO: Add Description}

\item{min_node_size_ave_prop}{TODO: Add Description}

\item{splitratio_first}{Proportion of the training data used as the splitting
dataset in the first stage.}

\item{splitratio_second}{Proportion of the training data used as the
splitting dataset in the second stage.}

\item{splitratio_prop}{TODO: Add Description}

\item{replace_first}{Sample with or without replacement in the first stage.}

\item{replace_second}{Sample with or without replacement in the second stage.}

\item{replace_prop}{TODO: Add Description}

\item{sample_fraction_first}{The size of total samples to draw for the
training data in the first stage.}

\item{sample_fraction_second}{The size of total samples to draw for the
training data in the second stage.}

\item{sample_fraction_prop}{TODO: Add Description}

\item{nthread}{Number of threads which should be used to work in parallel.}

\item{verbose}{TRUE for detailed output FALSE for no output}

\item{middleSplit_first}{TODO: Add Description}

\item{middleSplit_second}{TODO: Add Description}

\item{middleSplit_prop}{TODO: Add Description}
}
\description{
This is an implementation of the X-learner with honest random
forest in the first and second stage. The function returns an X-RF object.
}
