% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\docType{data}
\name{X_RF-X_RF}
\alias{X_RF-X_RF}
\alias{X_RF}
\alias{X_RF,}
\alias{X_RF-X_RF}
\title{X-Learner with honest RF for both stages}
\format{An object of class \code{NULL} of length 0.}
\usage{
X_RF(feat, tr, yobs, predmode = "propmean",
  relevant_Variable_first = 1:ncol(feat),
  relevant_Variable_second = 1:ncol(feat),
  relevant_Variable_prop = 1:ncol(feat), ntree_first = 1000,
  ntree_second = 1000, ntree_prop = 500, mtry_first = round(ncol(feat) *
  13/20), mtry_second = round(ncol(feat) * 17/20), mtry_prop = ncol(feat),
  min_node_size_spl_first = 2, min_node_size_ave_first = 1,
  min_node_size_spl_second = 5, min_node_size_ave_second = 6,
  min_node_size_spl_prop = 11, min_node_size_ave_prop = 33,
  splitratio_first = 1, splitratio_second = 0.8, splitratio_prop = 0.5,
  replace_first = TRUE, replace_second = TRUE, replace_prop = TRUE,
  sample_fraction_first = 0.8, sample_fraction_second = 0.7,
  sample_fraction_prop = 0.5, nthread = 0, verbose = TRUE,
  middleSplit_first = TRUE, middleSplit_second = TRUE,
  middleSplit_prop = FALSE)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{predmode}{One of propmean, control, treated, extreme. It specifies how
the two estimators of the second stage should be aggregated. The default is
propmean which refers to propensity score weighting.}

\item{relevant_Variable_first}{Variables which are only used in the first stage.}

\item{relevant_Variable_second}{Variables which are only used in the second stage.}

\item{ntree_first}{Numbers of trees in the first stage.}

\item{ntree_second}{Numbers of trees in the second stage.}

\item{mtry_first}{Numbers of trees in the second stage.}

\item{mtry_second}{Numbers of trees in the second stage.}

\item{min_node_size_spl_first}{minimum nodesize in the first stage for the
observations in the splitting set.}

\item{min_node_size_ave_first}{minimum nodesize in the first stage for the
observations in the average set.}

\item{min_node_size_spl_second}{minimum nodesize in the second stage for the
observations in the splitting set.}

\item{min_node_size_ave_second}{minimum nodesize in the second stage for the
observations in the averaging set.}

\item{splitratio_first}{Proportion of the training data used as the splitting
dataset in the first stage.}

\item{splitratio_second}{Proportion of the training data used as the
splitting dataset in the second stage.}

\item{replace_first}{Sample with or without replacement in the first stage.}

\item{replace_second}{Sample with or without replacement in the first stage.}

\item{sample_fraction_first}{The size of total samples to draw for the
training data in the first stage.}

\item{sample_fraction_second}{The size of total samples to draw for the
training data in the second stage.}

\item{nthread}{number of threats which should be used to work in parallel.}

\item{verbose}{whether or not to print messages of the training procedure.}
}
\value{
A `X_RF` object.
}
\description{
This is an implementation of the X-learner with honest random
forest in the first and second stage. The function returns an X-RF object.
}
\keyword{datasets}
