% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune_gpp.R
\name{X_RF_autotune_gpp}
\alias{X_RF_autotune_gpp}
\title{Autotuning for X-Learner with honest RF for both stages}
\usage{
X_RF_autotune_gpp(feat, tr, yobs, ntree = 500, nthread = 0,
  verbose = TRUE, init_points = 5, n_iter = 1, ...)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{ntree}{..}

\item{nthread}{..}

\item{Niter}{..}
}
\description{
DO NOT USE YET
}
