% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune.R
\docType{data}
\name{X_RF_autotune_hyperband}
\alias{X_RF_autotune_hyperband}
\title{Autotuning for X-Learner with honest RF for both stages}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{Niter}{number of iterations.}
}
\description{
This function tunes
}
\keyword{datasets}
