% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf_autotune_simple.R
\name{X_RF_autotune_simple}
\alias{X_RF_autotune_simple}
\title{Simple Autotuning for X-Learner with honest RF}
\usage{
X_RF_autotune_simple(feat, tr, yobs, ntree = 2000, ntree_testing = 600,
  nthread = 0, verbose = TRUE)
}
\arguments{
\item{feat}{A data frame of all the features.}

\item{tr}{A numeric vector contain 0 for control and 1 for treated variables.}

\item{yobs}{A numeric vector containing the observed outcomes.}

\item{ntree}{Number of trees used}

\item{ntree_testing}{...}

\item{nthread}{Number of threads which can run in parallel. If set 0, then
the maximum amount of possible threads is determined automatically. If set to
1 then the algorithm is absolutely deterministic (after specifying a seed).}

\item{verbose}{...}
}
\description{
This function tunes the X-Learner with honest random forest by
testing which of 11 prespecified settings seems to be the best
}
\examples{
  set.seed(14236142)
  feat <- iris[, -1]
  tr <- rbinom(nrow(iris), 1, .5)
  yobs <- iris[, 1]
  # train a
  xl_gpp <- X_RF_autotune_simple(feat, tr, yobs, ntree = 100, nthread = 0,
  verbose = FALSE)
  # computes the CATE and confidence intervals for CATE
  EstimateCate(xl_gpp, feat)
  CateCI(xl_gpp, feat, B = 5, verbose = FALSE)

}
\seealso{
\code{\link{X_RF_autotune_gpp}},
\code{\link{X_RF_autotune_hyperband}}
}
