% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhrf.R
\name{X_RF_fully_specified}
\alias{X_RF_fully_specified}
\title{X_RF_most_basic Constructor}
\usage{
X_RF_fully_specified(feat, tr, yobs, hyperparameter_list, verbose)
}
\arguments{
\item{feat}{feature data.frame.}

\item{tr}{treatment assignment 0 for control and 1 for treatment.}

\item{yobs}{the observed outcome.}

\item{hyperparameter_list}{A list of lists of hyper parameters}

\item{verbose}{TRUE for detailed output FALSE for no output}
}
\value{
A `X_RF` object.
}
\description{
This is the most basic X-learner with honest random forest
constructor. It should not be called by the user, since the list of
parameters is too big. Instead call the simpler version XhRF or one of the
self tuning versions
}
